+++ CHANGES.rst
@@ -12,6 +12,9 @@ Changes to API
 - Deprecate ``cast_arrays`` argument to ``from_fits_hdu`` and
   ``cast_fits_arrays`` argument to ``Datamodel.__init__`` [#214]
 
+- Use ``DataModel.__init__`` ``memmap`` argument when opening ASDF
+  files [#232]
+
 Other
 -----